requirements.txt
Bibliotecas necessárias para o projeto

supabase-py
requests
python-dotenv